user controller functions seperated out from model
